requirements.txt
altair==4.1.0
appnope==0.1.3
streamlit-drawable-canvas==0.9.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.5
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
coremltools==6.1
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
filelock==3.4.1
gitdb==4.0.9
GitPython==3.1.18
huggingface-hub==0.4.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
MarkupSafe==2.0.1
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numpy==1.19.5
opencv-python==4.5.1.48
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.27.1
rich==12.6.0
sacremoses==0.0.53
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.10.0
streamlit-drawable-canvas==0.9.2
sympy==1.9
terminado==0.12.1
testpath==0.6.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.12.0
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
transformers==4.15.0
typing_extensions==4.1.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zipp==3.6.0
